+++ pkg/plugin/queries.go
@@ -35,18 +35,27 @@ const (
 func buildFrames(responses []internal.FrameData, frameType frameType, frames []*data.Frame) []*data.Frame {
 	frameMap := make(map[string]*data.Frame)
 
+	var frameToAppendTo *data.Frame
 	for _, frameDataItem := range responses {
 		timestamp, err := frameDataItem.GetTimestamp()
 		if err != nil {
 			log.DefaultLogger.Error("error while parsing time %w", err)
 			continue
 		}
-		key := frameDataItem.GetKey()
-		frameToAppendTo, ok := frameMap[key]
-		if !ok {
-			frameDefinition := getFrameDefinitionFunction(frameType, frameDataItem)()
+		frameDefinition := getFrameDefinitionFunction(frameType, frameDataItem)()
+		// For table Wide frames, we always want to append to the one single frame in order
+		if frameType == TableFrameType && frameToAppendTo == nil {
 			frameToAppendTo = &frameDefinition
-			frameMap[key] = frameToAppendTo
+			frameMap["fixed-table"] = frameToAppendTo
+		} else if frameType == GraphFrameType {
+			key := frameDataItem.GetKey()
+
+			var ok bool
+			frameToAppendTo, ok = frameMap[key]
+			if !ok {
+				frameToAppendTo = &frameDefinition
+				frameMap[key] = frameToAppendTo
+			}
 		}
 
 		vals := getValDefinitionFunction(frameType, frameDataItem)(timestamp)
@@ -271,9 +280,11 @@ func QueryMonitorStatusPageChanges(ctx context.Context, query backend.DataQuery,
 			}
 			field.SetConfig(&data.FieldConfig{
 				Mappings: data.ValueMappings{
-					data.ValueMapper{"0": data.ValueMappingResult{Text: "(0) up", Color: "green"}},
-					data.ValueMapper{"1": data.ValueMappingResult{Text: "(1) degraded", Color: "yellow"}},
-					data.ValueMapper{"2": data.ValueMappingResult{Text: "(2) error", Color: "red"}},
+					data.ValueMapper{"0": data.ValueMappingResult{Text: "(0) unknown", Color: "gray"}},
+					data.ValueMapper{"1": data.ValueMappingResult{Text: "(1) maintenance", Color: "blue"}},
+					data.ValueMapper{"2": data.ValueMappingResult{Text: "(2) up", Color: "green"}},
+					data.ValueMapper{"3": data.ValueMappingResult{Text: "(3) degraded", Color: "yellow"}},
+					data.ValueMapper{"4": data.ValueMappingResult{Text: "(4) error", Color: "red"}},
 				},
 			})
 		}